Improve method listing in POD, and rewrite SYNOPSIS
The way the methods were listed in the documentation was a little
confusing; setting the title (head2) of each method as an example call
was not only different from CPAN standard, it was also hard to read in
the index. So, ``$do->droplet_create(\%args)`` becomes
``droplet_create``.

We also moved the example code up, so that it is the first thing in each
method documentation. Right below is the list of arguments, and a brief
description.